make Delta-I optional in infoclust
Lambert & Lance recommend basing merge order on the minimal increase
in information instead of selecting the merge with minimal level of
information. The resulting absolute information level is still used
in display and the order of merges is different from the levels
shown in trees. The argument for doing so is somewhat unclear to me, and the
resulting tree is not correctly handled by some hclust methods (including
cutree & rect.hclust).