Add iap: number of companion species and its significance
Old stray function from 2007 finds the average number of companion
species for each focal species and then generates numbers of the
same size as presences in the focal species to assess its
non-randomness.

The name and the idea comes from the "Index of Atmospheric
Purity" in lichen bioindication where this measure is used to
assess the bioindicator value of each lichen species.