add function for "peeled convex hulls"
in peeled convex hull we remove points from the margins of the hull
so that the resulting hull contains a given proportion of all points
in a small area. This serves in removing outliers. The function can
be plugged into vegan::ordihull(), but no work to this end was made.