Remove alt.kulczynski
The basic form of alternative Kulczynski is given as a/(b+c) in
Legendre & Legendre 7.15. Two identical sites have b = c = 0 and
hence similarity = Inf, and this cannot be turned into dissimilarity
(and is hardly meaningful). An inverse (b+c)/a would be Inf if there
are no shared species and hence that is outruled as well.