Mountford initial value index was given as a similarity
Also protect against -Inf and negative entries: The minimum
distance between two sites is obtained when J = min(A,B) and is
calculated as d = 1 - 2/|A-B|. This in d = -Inf if A = B and
d = -1 if |A-B| = 1.

NB., Mountford suggested this only as a way of getting initial values
in iterative estimation of similarity index I defined as
  exp(A*I) + exp(B*I) = exp((A+B-J)*I) + 1
and so calculated in vegdist (and given as distance log(2)-I).