add polar ordination w/manhattan, but warn against its use
McCune & Grace suggest this (p. 147-148), and it is easy to
implement. However, when you look at the graph, you think it is
Euclidean: this is not. This is so axis-centred that original Manhattan
distances can be recovered only if the axes are strictly parallel
to the columns of the original data, and this hardly ever is the case.