Educational toy-PC using species as axes
Selects the species that gives the best fit over all species when
used a single predictor variable. Then orthogonalizes the data against
the selected species, and repeat the search.

This is a kind of prelude to PCA: PCA finds the best linear combination
of species that gives the best fit over all species when used as single
predictor variable. This may be the same as Orloci's "position vector
ordination", but I got to go to the library to check the book that I
read some 30 years ago.

Also adds Gram-Schmidt orthogonalization as a separate function.