Orloci's position vectors were through sampling units
The implementation uses similar criteria as the previous code
(which was no renames spvectord), but instead of finding best
predictor species, it finds the sampling unit that maximizes the
vectors of other sampling unit when projected onto that vector.
The implementation reproduces the numerical example in Orloci's
chapter in Whittaker's "Ordination and Classification of Vegetation,
Handbook of Vegetation Science" (except that we scale the results
to variances).