find threshold of optimally split glm/gam response into two values
It is common in species distribution modelling to use tools that give
continuous prediction, but then urge to split these into binary outcome.
This function finds a threshold for optimal binary splitting in terms
of deviance. That is, the binary split is as close to the continuous
prediction as possible and gives the minimum deviance split. The two
response values will not be 0 and 1 in general, but something else, but
in terms of deviance, as different as possible. The function was developed
for binomial=1 (Bernoulli) models fitted with glm or mgcv::gam
functions.

This is only the skeleton of the function.