function to evaluate support for hclust groups with Bayes-Jaccard
First functions to implement a family of methods based on Bayesian
Jaccard index, where the dissimilarity is random variable from
Beta(b+c+1, a+1) in 2x2 notation. The expected value of this will
be (b+c+1)/(a+b+c+2) which is equal to Jaccard dissimilarity save
Bayes prior Beta(1, 1). The function will generate a number of
random samples from that Beta() and find how often each observed group
is found in classification of Beta() samples.