add a function to find mean distances of points
Mean distances are distances to the centroid of coordinates
generating the distances. Mean of distances does not define such
means. For instance, the mean of Euclidean distances in 2D plane
will not be in that plane, but in third and imaginary dimensions.
Math is Euclidean and based on Gower, but works for semimetrics
as well.