use ranked distances instead of raw distances: mix better
Constrained dissimilarities are typically shorter than community
dissimilarities, and when combined in MDSaddpoints, community points
will be expelled to the outskirts of ordination. Usink ranks of
dissimilarities for both allows better mixing. NB., in monoMDS of
one set (constrained, community) ranks of dissimilarities and
raw dissimilarities give the same result.